+++ src/main/kotlin/computer/obscure/twine/TwineError.kt
@@ -13,12 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.znci.twine.annotations
+package computer.obscure.twine
 
-/**
- * Annotation to mark a property as a native property in the Twine framework.
- * This allows the property to be exposed to Lua, making it accessible for manipulation from Lua scripts.
- */
-@Target(AnnotationTarget.PROPERTY)
-@Retention(AnnotationRetention.RUNTIME)
-annotation class TwineNativeProperty(val name: String = "INHERIT_FROM_DEFINITION")
+class TwineError(message: String) : Exception(message)
