+++ src/main/kotlin/computer/obscure/twine/TwineLuaValue.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.znci.twine
+package computer.obscure.twine
 
 import com.google.gson.JsonArray
 import com.google.gson.JsonElement
